requirements.txt
python-telegram-bot==20.7
openai==1.3.0
python-dotenv==1.0.0
prometheus-client==0.17.1
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-jaeger==1.18.0
opentelemetry-instrumentation==0.39b0
psutil==5.9.5
requests==2.31.0
google-generativeai==0.3.0  # ← ДОБАВЛЯЕМ ДЛЯ GEMINI
aiohttp==3.9.0  # ← ДОБАВЛЯЕМ для асинхронных HTTP запросов